requirements.txt
# MathStack - Stacked Optimization Proof of Concept
# Install into .venv: .venv\Scripts\pip install -r requirements.txt

chromadb>=0.4.0
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.0
optimum>=1.14.0
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
